GH Actions: add ini settings to "release" workflow
Follow-up on PR 65, which turned set `error_reporting` to `E_ALL` and turned `display_errors` on in the Test workflow.

Only just now noticed that I missed doing the same in the Release workflow.

This also changes the setting for `error_reporting` to `-1` as `E_ALL` does not contain **all** errors across PHP versions, while `-1` will show them, independently of the PHP version.

Also turning on assertions, just in case they are used under the hood in the test framework.